feat 당첨된 로또 금액과 구매한 로또 금액을 이용하여 수익률을 계산하는 기능

+++ src/main/java/lotto/domain/Money.java
@@ -35,7 +35,7 @@ public class Money {
             return true;
         }
 
-        public float calculateProfit(Map<Integer, Integer> resultLottos){
+        public long calculateProfit(Map<Integer, Integer> resultLottos){
             long percentage = MoneyCalculation.calculateResult(resultLottos) / money;
             return percentage;
         }
